plugins/Osmand-Traffic/py_fetch_traffic/pyproject.toml
[project]
name = "osmand-traffic-plugin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mercantile>=1.2.1",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "requests>=2.32.4",
]
